requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
matplotlib==3.8.3
pandas==2.2.2
python-decouple==3.8
seaborn==0.13.2
SQLAlchemy==2.0.27
Werkzeug==3.0.2
wordcloud==1.9.3